refactor(audio): promote windowing primitives into the task layer
extract_per_window_embeddings, its windowing grid and its Audio slicer move from
audio_analysis/embeddings.py into tasks/speaker_embeddings/windowing.py, and the
workflow imports them from there. Two callers now need them -- the workflow's
speaker-identity path and a future estimator -- and a task importing a workflow
would invert the dependency direction, since workflows compose tasks.

Promoted rather than duplicated: two copies of a windowing grid drift the
moment either is edited.

Behaviour, signatures and the 1.0/0.5 detection defaults are UNCHANGED; this is
a move, not a rewrite. quality.py's import of the two now-public helpers
(slice_audio, window_starts) is updated to the new names since it is the one
other in-repo consumer that used the old private spellings.

An AST guard now asserts nothing under audio/tasks imports audio/workflows,
with a companion test proving the guard can actually see a violation --
including one nested inside a function body, which a text search would miss.
The guard already failed before this move (speech_to_text_ensemble/api.py
imports audio_analysis.floors), a pre-existing violation unrelated to this
change and out of scope here; reported separately rather than silently fixed.
